requirements.txt
streamlit
pandas
plotly
scikit-learn
numpy
flask
flask-jwt-extended
flask-cors
bcrypt
python-dotenv
cryptography